Implement builder protocols for base structs
Implement builder protocols for base structs
---

* Duration
* Offset
* Multiplier
* Accidental
* PitchClass
* Pitch
* Notehead
* Note
* Rest
* Spacer
* Chord

Also
----

Don't trigger test watch on changing `.ly` and `.ily` files. Because
files of those types are created and removed during running the tests,
we can get into some gnarly situations where running tests trigger
running the again, and so on.

Update the versions of the `checkout` and `cache` actions in the
workflow to quiet some deprecation warnings